# docs/openapi/relay.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "AI模型接口",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
    "/v1/models": {
      "get": {
        "summary": "获取模型列表",
        "deprecated": false,
        "description": "获取当前可用的模型列表。\n\n根据请求头自动识别返回格式：\n- 包含 `x-api-key` 和 `anthropic-version` 头时返回 Anthropic 格式\n- 包含 `x-goog-api-key` 头或 `key` 查询参数时返回 Gemini 格式\n- 其他情况返回 OpenAI 格式\n",
        "operationId": "listModels",
        "tags": [
          "获取模型列表"
        ],
        "parameters": [
          {
            "name": "key",
            "in": "query",
            "description": "Google API Key (用于 Gemini 格式)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "description": "Anthropic API Key (用于 Claude 格式)",
            "required": false,
            "example": "",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "anthropic-version",
            "in": "header",
            "description": "Anthropic API 版本",
            "required": false,
            "example": "",
            "schema": {
              "type": "string",
              "example": "2023-06-01"
            }
          },
          {
            "name": "x-goog-api-key",
            "in": "header",
            "description": "Google API Key (用于 Gemini 格式)",
            "required": false,
            "example": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取模型列表",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelsResponse"
                }
              }
            },
            "headers": {}
          },
          "401": {
            "description": "认证失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v1beta/models": {
      "get": {
        "summary": "Gemini 格式获取",
        "deprecated": false,
        "description": "以 Gemini API 格式返回可用模型列表",
        "operationId": "listModelsGemini",
        "tags": [
          "获取模型列表"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "成功获取模型列表",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeminiModelsResponse"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v1/chat/completions": {
      "post": {
        "summary": "创建聊天对话",
        "deprecated": false,
        "description": "根据对话历史创建模型响应。支持流式和非流式响应。\n\n兼容 OpenAI Chat Completions API。\n",
        "operationId": "createChatCompletion",
        "tags": [
          "OpenAI格式(Chat)"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatCompletionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "成功创建响应",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatCompletionResponse"
                }
              }
            },
            "headers": {}
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          },
          "429": {
            "description": "请求频率限制",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v1/responses": {
      "post": {
        "summary": "创建响应 (OpenAI Responses API)",
        "deprecated": false,
        "description": "OpenAI Responses API，用于创建模型响应。\n支持多轮对话、工具调用、推理等功能。\n",
        "operationId": "createResponse",
        "tags": [
          "OpenAI格式(Responses)"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResponsesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "成功创建响应",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponsesResponse"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v1/responses/compact": {
      "post": {
        "summary": "压缩对话 (OpenAI Responses API)",
        "deprecated": false,
        "description": "OpenAI Responses API，用于对长对话进行 compaction。",
        "operationId": "compactResponse",
        "tags": [
          "OpenAI格式(Responses)"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResponsesCompactionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "成功压缩对话",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponsesCompactionResponse"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/v1/image-tasks/generations": {
      "post": {
        "summary": "Create an asynchronous image generation task",
        "description": "Accepts application/json and returns HTTP 202 immediately. The model defaults to gpt-image-2; when the effective model is gpt-image-2 and quality is omitted, quality defaults to high. Idempotency-Key and client_task_id may be used for idempotency; when both are supplied they must match. In-flight tasks are reused; successful terminal tasks are reused only while their result remains retrievable; failed terminal tasks are not reused. Idempotent replays that hit an existing task bypass new-task creation capacity and model request rate limits. Temporary request-storage capacity failures return HTTP 503 image_task_unavailable with Retry-After.",
        "operationId": "createAsyncImageGenerationTask",
        "tags": [
          "Image Workbench"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/PublicImageTaskIdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublicImageTaskCreateRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Task accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicImageTask"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/PublicImageTaskError"
          },
          "401": {
            "$ref": "#/components/responses/PublicImageTaskError"
          },
          "403": {
            "$ref": "#/components/responses/PublicImageTaskError"
          },
          "409": {
            "$ref": "#/components/responses/PublicImageTaskError"
          },
          "413": {
            "$ref": "#/components/responses/PublicImageTaskError"
          },
          "415": {
            "$ref": "#/components/responses/PublicImageTaskError"
          },
          "429": {
            "$ref": "#/components/responses/PublicImageTaskError"
          },
          "500": {
            "$ref": "#/components/responses/PublicImageTaskError"
          },
          "503": {
            "$ref": "#/components/responses/PublicImageTaskError"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v1/image-tasks/edits": {
      "post": {
        "summary": "Create an asynchronous image edit task",
        "description": "Accepts multipart/form-data with 1-6 image files under image, image[], or numeric indexed fields such as image[0] and image[1], then returns HTTP 202 immediately. Image presence and count are runtime-validated because OpenAPI 3.0 cannot model numeric dynamic multipart field names without rejecting valid image[n] uploads. The model defaults to gpt-image-2; when the effective model is gpt-image-2 and quality is omitted, quality defaults to high. Idempotency-Key and client_task_id may be used for idempotency; when both are supplied they must match. In-flight tasks are reused; successful terminal tasks are reused only while their result remains retrievable; failed terminal tasks are not reused. Idempotent replays that hit an existing task bypass new-task creation capacity and model request rate limits. Temporary request-storage capacity failures return HTTP 503 image_task_unavailable with Retry-After.",
        "operationId": "createAsyncImageEditTask",
        "tags": [
          "Image Workbench"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/PublicImageTaskIdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/PublicImageTaskEditRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Task accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicImageTask"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/PublicImageTaskError"
          },
          "401": {
            "$ref": "#/components/responses/PublicImageTaskError"
          },
          "403": {
            "$ref": "#/components/responses/PublicImageTaskError"
          },
          "409": {
            "$ref": "#/components/responses/PublicImageTaskError"
          },
          "413": {
            "$ref": "#/components/responses/PublicImageTaskError"
          },
          "415": {
            "$ref": "#/components/responses/PublicImageTaskError"
          },
          "429": {
            "$ref": "#/components/responses/PublicImageTaskError"
          },
          "500": {
            "$ref": "#/components/responses/PublicImageTaskError"
          },
          "503": {
            "$ref": "#/components/responses/PublicImageTaskError"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v1/image-tasks": {
      "get": {
        "summary": "Get multiple asynchronous image workbench tasks",
        "description": "Returns up to 100 tasks owned by the current user and current token. Image result data is not included. Requested IDs that are missing or not visible to this token are listed in not_found_ids without revealing why.",
        "operationId": "listAsyncImageTasks",
        "tags": [
          "Image Workbench"
        ],
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "required": true,
            "description": "Comma-separated task IDs",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Task statuses",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicImageTaskList"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/PublicImageTaskError"
          },
          "401": {
            "$ref": "#/components/responses/PublicImageTaskError"
          },
          "403": {
            "$ref": "#/components/responses/PublicImageTaskError"
          },
          "429": {
            "$ref": "#/components/responses/PublicImageTaskError"
          },
          "500": {
            "$ref": "#/components/responses/PublicImageTaskError"
          },
          "503": {
            "$ref": "#/components/responses/PublicImageTaskError"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v1/image-tasks/{task_id}": {
      "get": {
        "summary": "Get an asynchronous image workbench task",
        "description": "Returns status without image result data. The task must belong to the current user and current token.",
        "operationId": "getAsyncImageTask",
        "tags": [
          "Image Workbench"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/PublicImageTaskId"
          }
        ],
        "responses": {
          "200": {
            "description": "Task status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicImageTask"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/PublicImageTaskError"
          },
          "403": {
            "$ref": "#/components/responses/PublicImageTaskError"
          },
          "404": {
            "$ref": "#/components/responses/PublicImageTaskError"
          },
          "429": {
            "$ref": "#/components/responses/PublicImageTaskError"
          },
          "500": {
            "$ref": "#/components/responses/PublicImageTaskError"
          },
          "503": {
            "$ref": "#/components/responses/PublicImageTaskError"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v1/image-tasks/{task_id}/result": {
      "get": {
        "summary": "Get an asynchronous image workbench task result",
        "description": "Returns the NewAPI-retained OpenAI-compatible image response. New URL-only upstream image results are downloaded with SSRF protection, image-decoded, and cached as b64_json when the task result is stored; legacy URL results remain supported by the download endpoint. File-backed results are size/SHA-256 validated then streamed without loading the whole payload into memory. When known, responses include Content-Length and X-NewAPI-Result-SHA256 so clients can detect truncated transfers. Concurrent downloads are limited per process and per token (HTTP 429 rate_limit_exceeded when full). Unclaimed results expire no later than 72 hours after completion. Acknowledged results are unavailable after the two-minute grace period. HTTP 410 means the result is permanently gone; HTTP 503 means it is temporarily unreadable (for example while the shared result cache is unavailable) and the request should be retried.",
        "operationId": "getAsyncImageTaskResult",
        "tags": [
          "Image Workbench"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/PublicImageTaskId"
          },
          {
            "name": "access",
            "in": "query",
            "description": "Optional signed image-result capability. When supplied, this query value replaces Bearer authentication for this result request.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Original image result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/PublicImageTaskError"
          },
          "403": {
            "$ref": "#/components/responses/PublicImageTaskError"
          },
          "404": {
            "$ref": "#/components/responses/PublicImageTaskError"
          },
          "409": {
            "$ref": "#/components/responses/PublicImageTaskError"
          },
          "410": {
            "$ref": "#/components/responses/PublicImageTaskError"
          },
          "429": {
            "$ref": "#/components/responses/PublicImageTaskError"
          },
          "500": {
            "$ref": "#/components/responses/PublicImageTaskError"
          },
          "503": {
            "$ref": "#/components/responses/PublicImageTaskError"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v1/image-tasks/{task_id}/result/{image_index}/download": {
      "get": {
        "summary": "Download an asynchronous image workbench task result image",
        "description": "Downloads one image from the retained image task result through NewAPI as an attachment. This keeps browser downloads same-origin even when the upstream result used a cross-origin URL. The same authorization, settlement, retention and download concurrency rules as the result endpoint apply. The image_index parameter is zero-based.",
        "operationId": "downloadAsyncImageTaskResultImage",
        "tags": [
          "Image Workbench"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/PublicImageTaskId"
          },
          {
            "name": "image_index",
            "in": "path",
            "required": true,
            "description": "Zero-based image index inside the result data array.",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Image attachment",
            "headers": {
              "Content-Disposition": {
                "schema": {
                  "type": "string"
                },
                "description": "Attachment filename generated from task id and image index."
              },
              "X-Content-Type-Options": {
                "schema": {
                  "type": "string",
                  "example": "nosniff"
                }
              }
            },
            "content": {
              "image/*": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/PublicImageTaskError"
          },
          "401": {
            "$ref": "#/components/responses/PublicImageTaskError"
          },
          "403": {
            "$ref": "#/components/responses/PublicImageTaskError"
          },
          "404": {
            "$ref": "#/components/responses/PublicImageTaskError"
          },
          "409": {
            "$ref": "#/components/responses/PublicImageTaskError"
          },
          "410": {
            "$ref": "#/components/responses/PublicImageTaskError"
          },
          "429": {
            "$ref": "#/components/responses/PublicImageTaskError"
          },
          "500": {
            "$ref": "#/components/responses/PublicImageTaskError"
          },
          "503": {
            "$ref": "#/components/responses/PublicImageTaskError"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/v1/embeddings": {
      "post": {
        "summary": "创建文本嵌入",
        "deprecated": false,
        "description": "将文本转换为向量嵌入",
        "operationId": "createEmbedding",
        "tags": [
          "OpenAI格式(Embeddings)"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmbeddingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "成功创建嵌入",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmbeddingResponse"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ModelsResponse": {
        "type": "object",
        "properties": {
          "object": {
            "type": "string",
            "example": "list"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Model"
            }
          }
        }
      },
...
    },
...
    "parameters": {
      "PublicImageTaskId": {
        "name": "task_id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
      "PublicImageTaskIdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": false,
        "schema": {
          "type": "string",
          "maxLength": 191
        }
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "使用 Bearer Token 认证。\n格式: `Authorization: Bearer sk-xxxxxx`\n"
      },
      "SessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session",
        "description": "已废弃：Gin session 不再用于面板鉴权"
      },
      "AccessToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT or PAT",
        "description": "面板短期 JWT 或 User.AccessToken PAT；不再要求 New-Api-User"
      },
      "NewApiUser": {
        "type": "apiKey",
        "in": "header",
        "name": "New-Api-User",
        "description": "已废弃且不再参与鉴权；PAT 客户端可以移除此请求头"
      },
      "Combination": {
        "group": [
          {
            "id": "AccessToken"
          }
        ],
        "type": "combination"
      },
      "Combination1": {
        "group": [
          {
            "id": "AccessToken"
          }
        ],
        "type": "combination"
      },
      "Combination2": {
...
  },
  "servers": [],
  "security": [
    {
      "BearerAuth": []
    }
  ]
}