# router/relay-router.go
package router

import (
	"strings"

	"github.com/QuantumNous/new-api/constant"
	"github.com/QuantumNous/new-api/controller"
	"github.com/QuantumNous/new-api/middleware"
	"github.com/QuantumNous/new-api/relay"
	relayconstant "github.com/QuantumNous/new-api/relay/constant"
	"github.com/QuantumNous/new-api/types"

	"github.com/gin-gonic/gin"
)

func SetRelayRouter(router *gin.Engine) {
	router.Use(middleware.CORS())
	router.Use(middleware.DecompressRequestMiddleware())
	router.Use(middleware.BodyStorageCleanup()) // 清理请求体存储
	router.Use(middleware.StatsMiddleware())
	// https://platform.openai.com/docs/api-reference/introduction
	modelsRouter := router.Group("/v1/models")
	modelsRouter.Use(middleware.RouteTag("relay"))
	modelsRouter.Use(middleware.TokenAuth())
	{
		modelsRouter.GET("", func(c *gin.Context) {
			controller.ListModels(c, relayModelListChannelType(c))
		})

		modelsRouter.GET("/:model", func(c *gin.Context) {
			controller.RetrieveModel(c, relayModelRetrieveChannelType(c))
		})
	}

	geminiRouter := router.Group("/v1beta/models")
	geminiRouter.Use(middleware.RouteTag("relay"))
	geminiRouter.Use(middleware.TokenAuth())
	{
		geminiRouter.GET("", func(c *gin.Context) {
			controller.ListModels(c, constant.ChannelTypeGemini)
		})
	}

	geminiCompatibleRouter := router.Group("/v1beta/openai/models")
	geminiCompatibleRouter.Use(middleware.RouteTag("relay"))
	geminiCompatibleRouter.Use(middleware.TokenAuth())
	{
		geminiCompatibleRouter.GET("", func(c *gin.Context) {
			controller.ListModels(c, constant.ChannelTypeOpenAI)
		})
	}

	playgroundRouter := router.Group("/pg")
	playgroundRouter.Use(middleware.RouteTag("relay"))
	playgroundRouter.Use(middleware.SystemPerformanceCheck())
	playgroundRouter.Use(middleware.UserAuth(), middleware.Distribute())
	{
		playgroundRouter.POST("/chat/completions", controller.Playground)
	}
	relayV1Router := router.Group("/v1")
	relayV1Router.Use(middleware.RouteTag("relay"))
	relayV1Router.Use(middleware.SystemPerformanceCheck())
	relayV1Router.Use(middleware.TokenAuth())
	{
		// Responses WebSocket route. Channel selection happens after the first
		// response.create event; each event runs the ordinary request limiter.
		relayV1Router.GET("/responses", controller.ResponsesWebSocket)
	}
	relayV1Router.Use(middleware.ModelRequestRateLimit())
	imageTaskRouter := router.Group("/v1/image-tasks")
	imageTaskRouter.Use(middleware.RouteTag("relay"))
	imageTaskRouter.Use(middleware.SystemPerformanceCheck())
	imageTaskRouter.Use(middleware.RedactTaskArtifactAccessQuery())
	{
		imageTaskCreateRouter := imageTaskRouter.Group("")
		imageTaskCreateRouter.Use(middleware.TokenAuthForImageTaskCreation())
		imageTaskCreateRouter.POST("/generations",
			controller.RequirePublicImageTaskContentType(relayconstant.RelayModeImagesGenerations),
			controller.ReusePublicImageTaskGenerationIfExists,
			middleware.RejectExhaustedTokenForImageTaskCreation(),
			middleware.ImageTaskCreateAdmission(),
			middleware.ModelRequestRateLimit(),
			middleware.Distribute(),
			func(c *gin.Context) { controller.CreatePublicImageTask(c, relayconstant.RelayModeImagesGenerations) },
		)
		imageTaskCreateRouter.POST("/edits",
			controller.RequirePublicImageTaskContentType(relayconstant.RelayModeImagesEdits),
			controller.ReusePublicImageTaskEditIfExists,
			middleware.RejectExhaustedTokenForImageTaskCreation(),
			middleware.ImageTaskCreateAdmission(),
			middleware.ModelRequestRateLimit(),
			middleware.Distribute(),
			func(c *gin.Context) { controller.CreatePublicImageTask(c, relayconstant.RelayModeImagesEdits) },
		)

		imageTaskAccessRouter := imageTaskRouter.Group("")
		imageTaskAccessRouter.Use(middleware.TokenAuthForTaskAccess())
		imageTaskAccessRouter.Use(middleware.ImageTaskAccessRateLimit())
		imageTaskAccessRouter.GET("", controller.ListPublicImageTasks)
		imageTaskAccessRouter.GET("/:task_id", controller.GetPublicImageTask)
		imageTaskAccessRouter.GET("/:task_id/result/:image_index/download", controller.DownloadPublicImageTaskResultImage)
		imageTaskAccessRouter.POST("/:task_id/ack", controller.AcknowledgePublicImageTaskResult)
		imageTaskAccessRouter.POST("/:task_id/cancel", controller.CancelPublicImageTask)

		imageTaskResultRouter := imageTaskRouter.Group("")
		imageTaskResultRouter.Use(middleware.TokenAuthForImageTaskResultAccess())
		imageTaskResultRouter.Use(middleware.ImageTaskResultAccessRateLimit())
		imageTaskResultRouter.GET("/:task_id/result", controller.GetPublicImageTaskResult)
	}
	{
		// WebSocket 路由（统一到 Relay）
		wsRouter := relayV1Router.Group("")
		wsRouter.Use(middleware.Distribute())
		wsRouter.GET("/realtime", func(c *gin.Context) {
			controller.Relay(c, types.RelayFormatOpenAIRealtime)
		})
	}
	{
		//http router
		httpRouter := relayV1Router.Group("")
		httpRouter.Use(middleware.Distribute())

		// claude related routes
		// TODO: /messages/count_tokens is disabled. The current controller.CountClaudeTokens
		// httpRouter.POST("/messages/count_tokens", controller.CountClaudeTokens)
		httpRouter.POST("/messages", func(c *gin.Context) {
			controller.Relay(c, types.RelayFormatClaude)
		})

		// chat related routes
		httpRouter.POST("/completions", func(c *gin.Context) {
			controller.Relay(c, types.RelayFormatOpenAI)
		})
		httpRouter.POST("/chat/completions", func(c *gin.Context) {
			controller.Relay(c, types.RelayFormatOpenAI)
		})

		// response related routes: /v1/responses is registered by the
		// task-plugin protocol router so claimed models can enter the plugin
		// bridge and unclaimed models fall back to the normal Responses relay.
		// response related routes
		httpRouter.POST("/responses/compact", func(c *gin.Context) {
			controller.Relay(c, types.RelayFormatOpenAIResponsesCompaction)
		})

		// alpha search related routes (Codex standalone web search)
		httpRouter.POST("/alpha/search", func(c *gin.Context) {
			controller.Relay(c, types.RelayFormatOpenAIAlphaSearch)
		})

		// image related routes. /images/generations and /images/edits are
		// host protocol endpoints (openai_image) registered by
		// SetTaskPluginProtocolRouter; unclaimed models fall back to Relay.
		httpRouter.POST("/edits", func(c *gin.Context) {
			controller.Relay(c, types.RelayFormatOpenAIImage)
		})

		// embedding related routes
		httpRouter.POST("/embeddings", func(c *gin.Context) {
			controller.Relay(c, types.RelayFormatEmbedding)
		})

		// audio related routes
		httpRouter.POST("/audio/transcriptions", func(c *gin.Context) {
			controller.Relay(c, types.RelayFormatOpenAIAudio)
		})
		httpRouter.POST("/audio/translations", func(c *gin.Context) {
			controller.Relay(c, types.RelayFormatOpenAIAudio)
		})
		httpRouter.POST("/audio/speech", func(c *gin.Context) {
			controller.Relay(c, types.RelayFormatOpenAIAudio)
		})

		// rerank related routes
		httpRouter.POST("/rerank", func(c *gin.Context) {
			controller.Relay(c, types.RelayFormatRerank)
		})

		// gemini relay routes
		httpRouter.POST("/engines/:model/embeddings", func(c *gin.Context) {
			controller.Relay(c, types.RelayFormatGemini)
		})
		httpRouter.POST("/models/*path", func(c *gin.Context) {
			controller.Relay(c, types.RelayFormatGemini)
		})

		// other relay routes
		httpRouter.POST("/moderations", func(c *gin.Context) {
			controller.Relay(c, types.RelayFormatOpenAI)
		})

		// not implemented
		httpRouter.POST("/images/variations", controller.RelayNotImplemented)
		httpRouter.GET("/files", controller.RelayNotImplemented)
		httpRouter.POST("/files", controller.RelayNotImplemented)
		httpRouter.DELETE("/files/:id", controller.RelayNotImplemented)
		httpRouter.GET("/files/:id", controller.RelayNotImplemented)
		httpRouter.GET("/files/:id/content", controller.RelayNotImplemented)
		httpRouter.POST("/fine-tunes", controller.RelayNotImplemented)
		httpRouter.GET("/fine-tunes", controller.RelayNotImplemented)
		httpRouter.GET("/fine-tunes/:id", controller.RelayNotImplemented)
		httpRouter.POST("/fine-tunes/:id/cancel", controller.RelayNotImplemented)
		httpRouter.GET("/fine-tunes/:id/events", controller.RelayNotImplemented)
		httpRouter.DELETE("/models/:model", controller.RelayNotImplemented)
	}
	relayMjRouter := router.Group("/mj")
	relayMjRouter.Use(middleware.RouteTag("relay"))
	relayMjRouter.Use(middleware.SystemPerformanceCheck())
	registerMjRouterGroup(relayMjRouter)

	relayMjModeRouter := router.Group("/:mode/mj")
	relayMjModeRouter.Use(middleware.RouteTag("relay"))
	relayMjModeRouter.Use(middleware.SystemPerformanceCheck())
	registerMjRouterGroup(relayMjModeRouter)
	//relayMjRouter.Use()

	relaySunoRouter := router.Group("/suno")
	relaySunoRouter.Use(middleware.RouteTag("relay"))
	relaySunoRouter.Use(middleware.SystemPerformanceCheck())
	{
		relaySunoRouter.POST("/submit/:action", middleware.TokenAuth(), middleware.Distribute(), controller.RelayTask)
		relaySunoRouter.POST("/fetch", middleware.TokenAuthAllowExhausted(), middleware.Distribute(), controller.RelayTaskFetch)
		relaySunoRouter.GET("/fetch/:id", middleware.TokenAuthAllowExhausted(), middleware.Distribute(), controller.RelayTaskFetch)
	}
	relayGeminiRouter := router.Group("/v1beta")
	// :countTokens is not implemented. Answer it like an unregistered route
	// before auth/channel selection instead of silently relaying it as
	// generateContent (#7283).
	relayGeminiRouter.Use(func(c *gin.Context) {
		if strings.HasSuffix(c.Request.URL.Path, ":countTokens") {
			controller.RelayNotFound(c)
			c.Abort()
		}
	})
	relayGeminiRouter.Use(middleware.RouteTag("relay"))
	relayGeminiRouter.Use(middleware.SystemPerformanceCheck())
	relayGeminiRouter.Use(middleware.TokenAuth())
	relayGeminiRouter.Use(middleware.ModelRequestRateLimit())
	relayGeminiRouter.Use(middleware.Distribute())
	{
		// Gemini API 路径格式: /v1beta/models/{model_name}:{action}
		relayGeminiRouter.POST("/models/*path", func(c *gin.Context) {
			controller.Relay(c, types.RelayFormatGemini)
		})
	}
}

func relayModelListChannelType(c *gin.Context) int {
	switch {
	case c.GetHeader("x-api-key") != "" && c.GetHeader("anthropic-version") != "":
		return constant.ChannelTypeAnthropic
	case c.GetHeader("x-goog-api-key") != "" || c.Query("key") != "":
		return constant.ChannelTypeGemini
	default:
		return constant.ChannelTypeOpenAI
	}
}

func relayModelRetrieveChannelType(c *gin.Context) int {
	switch {
	case c.GetHeader("x-api-key") != "" && c.GetHeader("anthropic-version") != "":
		return constant.ChannelTypeAnthropic
	case c.GetHeader("x-goog-api-key") != "" || c.Query("key") != "":
		return constant.ChannelTypeGemini
	default:
		return constant.ChannelTypeOpenAI
	}
}

func registerMjRouterGroup(relayMjRouter *gin.RouterGroup) {
	relayMjRouter.GET("/image/:id", middleware.RedactTaskArtifactAccessQuery(), relay.RelayMidjourneyImage)
	mjReadRouter := relayMjRouter.Group("")
	mjReadRouter.Use(middleware.TokenAuthAllowExhausted(), middleware.Distribute())
	{
		mjReadRouter.GET("/task/:id/fetch", controller.RelayMidjourney)
		mjReadRouter.GET("/task/:id/image-seed", controller.RelayMidjourney)
		mjReadRouter.POST("/task/list-by-condition", controller.RelayMidjourney)
	}
	mjWriteRouter := relayMjRouter.Group("")
	mjWriteRouter.Use(middleware.TokenAuth(), middleware.Distribute())
	{
		mjWriteRouter.POST("/submit/action", controller.RelayMidjourney)
		mjWriteRouter.POST("/submit/shorten", controller.RelayMidjourney)
		mjWriteRouter.POST("/submit/modal", controller.RelayMidjourney)
		mjWriteRouter.POST("/submit/imagine", controller.RelayMidjourney)
		mjWriteRouter.POST("/submit/change", controller.RelayMidjourney)
		mjWriteRouter.POST("/submit/simple-change", controller.RelayMidjourney)
		mjWriteRouter.POST("/submit/describe", controller.RelayMidjourney)
		mjWriteRouter.POST("/submit/blend", controller.RelayMidjourney)
		mjWriteRouter.POST("/submit/edits", controller.RelayMidjourney)
		mjWriteRouter.POST("/submit/video", controller.RelayMidjourney)
		mjWriteRouter.POST("/insight-face/swap", controller.RelayMidjourney)
		mjWriteRouter.POST("/submit/upload-discord-images", controller.RelayMidjourney)
	}
}